Add prepend methods to pure python impl
