+++ python-impl/README.md
@@ -1,6 +1,6 @@
 ### BLS12-381 and Signatures in python
 
 Implements the BLS12 curve and optimal ate pairing, as well
-as BLS signatures and aggregation. Use for reference / educational purposes only.
+as BLS signatures and aggregation. Use for reference / educational purposes only. For an optimized implementation, use the [Python bindings](https://github.com/Chia-Network/bls-signatures/tree/master/python-bindings).
 
 For a good introduction to pairings, read [Pairings for Beginners](http://www.craigcostello.com.au/pairings/PairingsForBeginners.pdf) by Craig Costello.
